src/requirements.txt
Flask
requests
python-dotenv
mysql-connector-python

